Removing analytics_support, and moving the helper methods into the pipelines helper
