--- server/webapp/WEB-INF/rails.new/lib/analytics_support.rb
@@ -1,29 +1,0 @@
-##########################################################################
-# Copyright 2017 ThoughtWorks, Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-##########################################################################
-
-module AnalyticsSupport
-  include JavaImports
-
-  def with_pipeline_analytics_support(&block)
-    return unless block_given?
-
-    default_plugin_info_finder.allPluginInfos(PluginConstants.ANALYTICS_EXTENSION).each do |plugin|
-      if plugin.getCapabilities().supportsPipelineAnalytics()
-        yield plugin.getDescriptor().id()
-      end
-    end
-  end
-end
